requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
jinja2
pymysql
cryptography
pydantic-settings
psutil
